+++ src/main/java/com/github/johnmedlockdev/main/modes/interfaces/ModeInterface.java
@@ -6,12 +6,13 @@ import java.util.Scanner;
 
 public interface ModeInterface {
 
-    void gatherInput();
-    static void startOver(){
+    static void startOver() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Select Mode: Manual : M || Batch : B || Generate : G || Advise : A || Exit : E");
         String[] input = {scanner.nextLine()};
         new Init(input);
         scanner.close();
     }
+
+    void gatherInput();
 }