+++ src/test/java/com/github/johnmedlockdev/main/database/DatabaseTest.java
@@ -1,0 +1,15 @@
+package com.github.johnmedlockdev.main.database;
+
+import org.junit.Test;
+
+public class DatabaseTest {
+    @Test
+    public void selectStmts() {
+        Database db = new Database("Select", "TEST");
+    }
+
+    @Test
+    public void insertStmts() {
+        Database db = new Database("insert", "TEST");
+    }
+}