+++ src/main/java/com/github/johnmedlockdev/main/Main.java
@@ -2,6 +2,7 @@ package com.github.johnmedlockdev.main;
 
 import com.github.johnmedlockdev.main.routes.Init;
 
+
 public class Main {
     public static void main(String... args) {
         new Init(args);